Invalid array size check
there was a problem with the way property related to array size check was passed. Not only was the condition inverse, but also the properties weren't passed.

With this change those issues were fixed, we've added the missing maven plugin feature and ensured that the array check is the default option.

fixes #60