Unable to download stubs from repository via HTTPS
without this change you can't specify the location of your private key. also without this change it's impossible to override the aether's behaviour in a meaningful way.
with this change we're adding two properties to pass the private key's location and passwod. Also the aether's repo setup logic is customizable.

fixes #195