gh-245: Updated to reflect pull request feedback
Changed the naming of methods through to the RegEx patterns to match the suggestions given in the pull request.  Added a new anyNonEmptyString which was suggested to better reflect utility style methods that are present in most libraries, specifically apache commons in this case, which has notEmpty and notBlank.  What used to be anyNotEmptyOrWhitespaceString is now anyNonBlankString.

Refactored the JaxRsClientMethodBuilderSpec and MockMvcMethodBodyBuilderSpec to use a data table value that reflect the escaped and unescaped regex $ symbol in order to get rid of the spaghetti code assertions that were there