Fix property source ordering issue with bootstrap context
The problem is that the bootstrap context gets created first
with the correct wiremock server port, and then the environment
is merged into the main context before the annotation
property source is created, so the latter takes precedence.
Fixed by re-ordering the property sources if detected.

Also switch off bootstrap context for wiremock server

Fixes gh-225