Cannot set fixed port in mvn stub runner (httpPort is ignored)
without this change the value of `spring.cloud.contract.verifier.http.port` gets executed via the `httpPort` method in options builder. The problem is that this method applies the port to the previously started stub. In this case we're not providing any stubs explicitly thus things don't work as they should.

with this change we reuse the min / max port mechanism to limit the range of port to the provided http port

fixes #384