Inconsistent regex() usage and invalid client stubs after passed server verification
After a short discussion, we came to the conclusion that it's confusing for RegexPatterns to return String. When defining the consumer/producer pair, you can easily create by mistake 2 string values if you forget the regex(...) and then you will only notice the issues on the consumer side due to invalid regular expression.

So the solution would be to change the current return types to Pattern. The regex(String) method will remain but also we need to add regex(Pattern) to maintain the compatibility of current DSLs. regex(Pattern) will return just the Pattern.

After such a change failing fast will be much easier.

fixes #389