`web` dependency is optional, `StubRunner` listener conditional
without this change we always include `spring-boot-starter-web` to Stub Runner. With this change the dependency is optional. Thanks to this reactive applications will be able to work fine with contract on the classpath.

without this change we always run Stub Runner test execution listener, even if the test class is not annotated. With this change we run the listener only, if the test class was annotated with `@AutoConfigureStubRunner`.

Also we add information about the `java.net.SocketException: Unexpected end of file from server` exception to the docs and how to override that issue.

Fixes gh-798, gh-809, gh-799