Added support for the stubs:// protocol
Currently Stub Runner can fetch a JAR, unpack it and only then pick the contract and stubs. Sometimes, what you want to achieve, is to point to a given folder where the stubs are already unpacked. Another option is a multimodule project, where the stubs are generated and then the user would like to just reference those instead of installing a JAR in a local m2.

With this change we suggest adding a protocol called stubs that would allow to define a path where the unpacked contracts and stub lay.

fixes gh-1150