Added the option "inProgress" to the contract
without this change we can't have a contract that will generate a stub but not a test. Of course that's for a reason since it's overriding the very essence of contract test.
with this change we assume that the users know what they're doing and that they will use this flag rarerly. Cause it means that you can have a false positive.

related to gh-881