Disable broken incremental build for InitContractsTask (#1186)
With current implementation of the `GradleContractsDownloader`, where
we have custom caching, `InitContractsTask` doesn't really need to be
incremental. Moreover, there is an issue in current implementation of
the `InitContractsTask`: it doesn't mark itself as out of date when
dynamic dependency is used (`+` or `SNAPSHOT` version). So for now just
disabling incremental build for that task.

Also restored old behaviour, where we used to always use `".*"` pattern
to define included contracts for `GenerateServerTests` task.

Also fixed minor logging issue.

Fixes gh-1133