Update integration test to take advantage of java-gradle-plugin
This plugin is required when using the withPluginClasspath(..) method,
but during Gradle execution it creates a listing of the classpath that
the plugin under test would have. We can then use that explicitly to
build up the necessary classpath in the init.gradle script for loading
the necessary classes to run integration tests against the plugin
without needing to copy all of the jar files into a folder via a Gradle
task.