Messaging support for stub runner boot
With this change by adding --thin.profile=kafka or --thin.profile=rabbit we're able to turn on messaging support for stub runner boot.

Also for Docker, you will be able to pass MESSAGING_TYPE env var with either rabbit or kafka as a value and we'll run the messaging support for you.

fixes gh-1468