Select value of an `anyOf()` property at random
The current implementation of `anyOf()` always only selects the first value
when generating a property. It means that other possible enum values are always
ignored when producing a value for the property.

This changes the behaviour so that the value is picked at random from any of
the provided values. It could be considered a breaking changes because of the
change of behaviour although it shouldn't be from the contract perspective.

Fixes #1424